Include the option to write the message found
When the player reaches the end of the labyrinth, now a congratulation
window appears together with a space where the message can be written.
If the message is correctly written, the minigame finishes.

I have included this changes also on the MainFile, and I had to create
an experiments file to test what I found on the internet. I also
included another image to the data folder.