Improve the series minigame
I have included the image with the series on the Series minigame,
and the spaces where you can set the correct value for the series
I have also included the alphabet, so that the player can get the
message, however it is still not working as I want to.
write the number that follows the series. I have also included the
alphabet, however it is still not working as I want to.