Finish the crossword and learn packages
I finished the crossword game and now it is included in the main file.

I tried to learn how to create a sliding puzzle but I disregarded that
game because it would have been too complex for my cousin's daughter.

I started learning the turtle module because I saw that I can do screen
interaction with that module and it looks quite feasible. I made a
tutorial on that module, and once I felt confident, I started creating
the circuit of pipelines which would be the background of the main game.